+++ x/perp/keeper/hooks.go
@@ -3,6 +3,9 @@ package keeper
 import (
 	"time"
 
+	"github.com/NibiruChain/nibiru/collections/keys"
+	"github.com/NibiruChain/nibiru/x/common"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	epochstypes "github.com/NibiruChain/nibiru/x/epochs/types"
@@ -18,7 +21,7 @@ func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, _ int64) 
 		return
 	}
 
-	for _, pairMetadata := range k.PairMetadataState(ctx).GetAll() {
+	for _, pairMetadata := range k.PairsMetadata.Iterate(ctx, keys.NewRange[common.AssetPair]()).Values() {
 		if !k.VpoolKeeper.ExistsPool(ctx, pairMetadata.Pair) {
 			ctx.Logger().Error("no pool for pair found", "pairMetadata.Pair", pairMetadata.Pair)
 			continue
@@ -55,7 +58,7 @@ func (k Keeper) AfterEpochEnd(ctx sdk.Context, epochIdentifier string, _ int64) 
 		}
 
 		pairMetadata.CumulativeFundingRates = append(pairMetadata.CumulativeFundingRates, cumulativeFundingRate)
-		k.PairMetadataState(ctx).Set(pairMetadata)
+		k.PairsMetadata.Insert(ctx, pairMetadata.Pair, pairMetadata)
 
 		if err = ctx.EventManager().EmitTypedEvent(&types.FundingRateChangedEvent{
 			Pair:                  pairMetadata.Pair.String(),