+++ x/perp/keeper/liquidate.go
@@ -3,6 +3,8 @@ package keeper
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
+	"github.com/NibiruChain/nibiru/collections/keys"
+
 	"github.com/NibiruChain/nibiru/x/common"
 	"github.com/NibiruChain/nibiru/x/perp/types"
 	vpooltypes "github.com/NibiruChain/nibiru/x/vpool/types"
@@ -34,14 +36,14 @@ func (k Keeper) Liquidate(
 		return sdk.Coin{}, sdk.Coin{}, err
 	}
 
-	position, err := k.PositionsState(ctx).Get(pair, traderAddr)
+	position, err := k.Positions.Get(ctx, keys.Join(pair, keys.String(traderAddr.String())))
 	if err != nil {
 		return sdk.Coin{}, sdk.Coin{}, err
 	}
 
 	marginRatio, err := k.GetMarginRatio(
 		ctx,
-		*position,
+		position,
 		types.MarginCalculationPriceOption_MAX_PNL,
 	)
 	if err != nil {
@@ -50,7 +52,7 @@ func (k Keeper) Liquidate(
 
 	if k.VpoolKeeper.IsOverSpreadLimit(ctx, pair) {
 		marginRatioBasedOnOracle, err := k.GetMarginRatio(
-			ctx, *position, types.MarginCalculationPriceOption_INDEX)
+			ctx, position, types.MarginCalculationPriceOption_INDEX)
 		if err != nil {
 			return sdk.Coin{}, sdk.Coin{}, err
 		}
@@ -67,16 +69,16 @@ func (k Keeper) Liquidate(
 	}
 
 	marginRatioBasedOnSpot, err := k.GetMarginRatio(
-		ctx, *position, types.MarginCalculationPriceOption_SPOT)
+		ctx, position, types.MarginCalculationPriceOption_SPOT)
 	if err != nil {
 		return sdk.Coin{}, sdk.Coin{}, err
 	}
 
 	var liquidationResponse types.LiquidateResp
 	if marginRatioBasedOnSpot.GTE(params.LiquidationFeeRatio) {
-		liquidationResponse, err = k.ExecutePartialLiquidation(ctx, liquidatorAddr, position)
+		liquidationResponse, err = k.ExecutePartialLiquidation(ctx, liquidatorAddr, &position)
 	} else {
-		liquidationResponse, err = k.ExecuteFullLiquidation(ctx, liquidatorAddr, position)
+		liquidationResponse, err = k.ExecuteFullLiquidation(ctx, liquidatorAddr, &position)
 	}
 	if err != nil {
 		return sdk.Coin{}, sdk.Coin{}, err
@@ -297,7 +299,7 @@ func (k Keeper) ExecutePartialLiquidation(
 		Mul(params.LiquidationFeeRatio)
 	positionResp.Position.Margin = positionResp.Position.Margin.
 		Sub(liquidationFeeAmount)
-	k.PositionsState(ctx).Set(positionResp.Position)
+	k.Positions.Insert(ctx, keys.Join(positionResp.Position.Pair, keys.String(positionResp.Position.TraderAddress)), *positionResp.Position)
 
 	// Compute splits for the liquidation fee
 	feeToLiquidator := liquidationFeeAmount.QuoInt64(2)