+++ x/perp/keeper/liquidate_unit_test.go
@@ -5,6 +5,9 @@ import (
 	"testing"
 	"time"
 
+	"github.com/NibiruChain/nibiru/collections"
+	"github.com/NibiruChain/nibiru/collections/keys"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/stretchr/testify/assert"
@@ -93,14 +96,14 @@ func TestLiquidateIntoPartialLiquidation(t *testing.T) {
 				Margin:        tc.initialPositionMargin,
 				OpenNotional:  tc.initialPositionOpenNotional,
 			}
-			perpKeeper.PositionsState(ctx).Set(&position)
+			setPosition(perpKeeper, ctx, position)
 
 			t.Log("set params")
 			params := types.DefaultParams()
 			perpKeeper.SetParams(ctx, params)
 
 			t.Log("set pair metadata")
-			perpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(perpKeeper, ctx, types.PairMetadata{
 				Pair: common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{
 					sdk.ZeroDec(),
@@ -180,7 +183,7 @@ func TestLiquidateIntoPartialLiquidation(t *testing.T) {
 			assert.EqualValues(t, tc.expectedPerpEFFee, feeToFund)
 
 			t.Log("assert new position and event")
-			newPosition, err := perpKeeper.PositionsState(ctx).Get(common.Pair_BTC_NUSD, traderAddr)
+			newPosition, err := perpKeeper.Positions.Get(ctx, keys.Join(common.Pair_BTC_NUSD, keys.String(traderAddr.String())))
 			require.NoError(t, err)
 			assert.EqualValues(t, traderAddr.String(), newPosition.TraderAddress)
 			assert.EqualValues(t, common.Pair_BTC_NUSD, newPosition.Pair)
@@ -266,14 +269,14 @@ func TestLiquidateIntoFullLiquidation(t *testing.T) {
 				Margin:        tc.initialPositionMargin,
 				OpenNotional:  tc.initialPositionOpenNotional,
 			}
-			perpKeeper.PositionsState(ctx).Set(&position)
+			setPosition(perpKeeper, ctx, position)
 
 			t.Log("set params")
 			params := types.DefaultParams()
 			perpKeeper.SetParams(ctx, params)
 
 			t.Log("set pair metadata")
-			perpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(perpKeeper, ctx, types.PairMetadata{
 				Pair: common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{
 					sdk.ZeroDec(),
@@ -346,9 +349,9 @@ func TestLiquidateIntoFullLiquidation(t *testing.T) {
 			assert.EqualValues(t, tc.expectedPerpEFFee.String(), feeToFund.String())
 
 			t.Log("assert new position and event")
-			newPosition, err := perpKeeper.PositionsState(ctx).Get(common.Pair_BTC_NUSD, traderAddr)
-			require.ErrorIs(t, err, types.ErrPositionNotFound)
-			assert.Nil(t, newPosition)
+			newPosition, err := perpKeeper.Positions.Get(ctx, keys.Join(common.Pair_BTC_NUSD, keys.String(traderAddr.String())))
+			require.ErrorIs(t, err, collections.ErrNotFound)
+			assert.Empty(t, newPosition)
 
 			testutilevents.RequireHasTypedEvent(t, ctx, &types.PositionLiquidatedEvent{
 				Pair:                  common.Pair_BTC_NUSD.String(),
@@ -435,14 +438,14 @@ func TestLiquidateIntoFullLiquidationWithBadDebt(t *testing.T) {
 				Margin:        tc.initialPositionMargin,
 				OpenNotional:  tc.initialPositionOpenNotional,
 			}
-			perpKeeper.PositionsState(ctx).Set(&position)
+			setPosition(perpKeeper, ctx, position)
 
 			t.Log("set params")
 			params := types.DefaultParams()
 			perpKeeper.SetParams(ctx, params)
 
 			t.Log("set pair metadata")
-			perpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(perpKeeper, ctx, types.PairMetadata{
 				Pair: common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{
 					sdk.ZeroDec(),
@@ -518,9 +521,9 @@ func TestLiquidateIntoFullLiquidationWithBadDebt(t *testing.T) {
 			assert.EqualValues(t, tc.expectedPerpEFFee.String(), feeToFund.String())
 
 			t.Log("assert new position and event")
-			newPosition, err := perpKeeper.PositionsState(ctx).Get(common.Pair_BTC_NUSD, traderAddr)
-			require.ErrorIs(t, err, types.ErrPositionNotFound)
-			assert.Nil(t, newPosition)
+			newPosition, err := perpKeeper.Positions.Get(ctx, keys.Join(common.Pair_BTC_NUSD, keys.String(traderAddr.String())))
+			require.ErrorIs(t, err, collections.ErrNotFound)
+			assert.Empty(t, newPosition)
 
 			testutilevents.RequireHasTypedEvent(t, ctx, &types.PositionLiquidatedEvent{
 				Pair:                  common.Pair_BTC_NUSD.String(),
@@ -919,7 +922,7 @@ func TestKeeper_ExecuteFullLiquidation(t *testing.T) {
 			newParams := types.DefaultParams()
 			newParams.LiquidationFeeRatio = tc.liquidationFee
 			perpKeeper.SetParams(ctx, newParams)
-			perpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(perpKeeper, ctx, types.PairMetadata{
 				Pair: common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{
 					sdk.ZeroDec(), // zero funding payment for this test case
@@ -959,7 +962,7 @@ func TestKeeper_ExecuteFullLiquidation(t *testing.T) {
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    ctx.BlockHeight(),
 			}
-			perpKeeper.PositionsState(ctx).Set(&position)
+			setPosition(perpKeeper, ctx, position)
 
 			t.Log("execute full liquidation")
 			liquidationResp, err := perpKeeper.ExecuteFullLiquidation(