+++ x/perp/keeper/msg_server_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/NibiruChain/nibiru/collections"
+
 	simapp2 "github.com/NibiruChain/nibiru/simapp"
 
 	"github.com/cosmos/cosmos-sdk/simapp"
@@ -48,7 +50,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 			traderFunds:     sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 1000)),
 			initialPosition: nil,
 			margin:          sdk.NewInt64Coin(common.DenomNUSD, 1000),
-			expectedErr:     types.ErrPositionNotFound,
+			expectedErr:     collections.ErrNotFound,
 		},
 		{
 			name:        "success",
@@ -85,7 +87,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
-			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(app.PerpKeeper, ctx, types.PairMetadata{
 				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
@@ -96,7 +98,7 @@ func TestMsgServerAddMargin(t *testing.T) {
 			if tc.initialPosition != nil {
 				t.Log("create position")
 				tc.initialPosition.TraderAddress = traderAddr.String()
-				require.NoError(t, app.PerpKeeper.PositionsState(ctx).Create(tc.initialPosition))
+				setPosition(app.PerpKeeper, ctx, *tc.initialPosition)
 			}
 
 			resp, err := msgServer.AddMargin(sdk.WrapSDKContext(ctx), &types.MsgAddMargin{
@@ -153,7 +155,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 			vaultFunds:      sdk.NewCoins(sdk.NewInt64Coin(common.DenomNUSD, 0)),
 			initialPosition: nil,
 			marginToRemove:  sdk.NewInt64Coin(common.DenomNUSD, 1000),
-			expectedErr:     types.ErrPositionNotFound,
+			expectedErr:     collections.ErrNotFound,
 		},
 		{
 			name:       "vault insufficient funds",
@@ -204,7 +206,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
-			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(app.PerpKeeper, ctx, types.PairMetadata{
 				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
@@ -215,7 +217,7 @@ func TestMsgServerRemoveMargin(t *testing.T) {
 			if tc.initialPosition != nil {
 				t.Log("create position")
 				tc.initialPosition.TraderAddress = traderAddr.String()
-				require.NoError(t, app.PerpKeeper.PositionsState(ctx).Create(tc.initialPosition))
+				setPosition(app.PerpKeeper, ctx, *tc.initialPosition)
 			}
 
 			ctx = ctx.WithBlockTime(ctx.BlockTime().Add(time.Second * 5)).WithBlockHeight(ctx.BlockHeight() + 1)
@@ -290,7 +292,7 @@ func TestMsgServerOpenPosition(t *testing.T) {
 				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
-			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(app.PerpKeeper, ctx, types.PairMetadata{
 				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
@@ -375,13 +377,13 @@ func TestMsgServerClosePosition(t *testing.T) {
 				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
-			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(app.PerpKeeper, ctx, types.PairMetadata{
 				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
 
 			t.Log("create position")
-			require.NoError(t, app.PerpKeeper.PositionsState(ctx).Create(&types.Position{
+			setPosition(app.PerpKeeper, ctx, types.Position{
 				TraderAddress:                  tc.traderAddr.String(),
 				Pair:                           tc.pair,
 				Size_:                          sdk.OneDec(),
@@ -389,7 +391,7 @@ func TestMsgServerClosePosition(t *testing.T) {
 				OpenNotional:                   sdk.OneDec(),
 				LatestCumulativeFundingPayment: sdk.ZeroDec(),
 				BlockNumber:                    1,
-			}))
+			})
 			require.NoError(t, simapp.FundModuleAccount(app.BankKeeper, ctx, types.VaultModuleAccount, sdk.NewCoins(sdk.NewInt64Coin(tc.pair.QuoteDenom(), 1))))
 
 			resp, err := msgServer.ClosePosition(sdk.WrapSDKContext(ctx), &types.MsgClosePosition{
@@ -472,7 +474,7 @@ func TestMsgServerLiquidate(t *testing.T) {
 				/* maintenanceMarginRatio */ sdk.MustNewDecFromStr("0.0625"),
 				/* maxLeverage */ sdk.MustNewDecFromStr("15"),
 			)
-			app.PerpKeeper.PairMetadataState(ctx).Set(&types.PairMetadata{
+			setPairMetadata(app.PerpKeeper, ctx, types.PairMetadata{
 				Pair:                   common.Pair_BTC_NUSD,
 				CumulativeFundingRates: []sdk.Dec{sdk.ZeroDec()},
 			})
@@ -488,7 +490,7 @@ func TestMsgServerLiquidate(t *testing.T) {
 				require.NoError(t, app.PricefeedKeeper.GatherRawPrices(ctx, pair.BaseDenom(), pair.QuoteDenom()))
 
 				t.Log("create position")
-				require.NoError(t, app.PerpKeeper.PositionsState(ctx).Create(&types.Position{
+				setPosition(app.PerpKeeper, ctx, types.Position{
 					TraderAddress:                  traderAddr.String(),
 					Pair:                           pair,
 					Size_:                          sdk.OneDec(),
@@ -496,7 +498,7 @@ func TestMsgServerLiquidate(t *testing.T) {
 					OpenNotional:                   sdk.NewDec(2), // new spot price is 1, so position can be liquidated
 					LatestCumulativeFundingPayment: sdk.ZeroDec(),
 					BlockNumber:                    1,
-				}))
+				})
 				require.NoError(t, simapp.FundModuleAccount(app.BankKeeper, ctx, types.VaultModuleAccount, sdk.NewCoins(sdk.NewInt64Coin(pair.QuoteDenom(), 1))))
 			}